feat(dto): :sparkles: added new dto, model and value objects for book registration
Extended the library domain to include a BookRegistrationDto along with corresponding models for Author, and value objects for Category and Isbn. This allows for structured data transfer when processing new book registrations. Also, updated the book registration service by changing validator and method parameter types to use this new dto.