feat(BookRegistration): :sparkles: enhanced book registration function
The book registration process in the library domain service has been enhanced. Now, the method `registerBook` instead of returning an Object now returns an instance of Book after validating the incoming data and persisting the new book to the repository. This refinement makes the function more specific and coherent and delegates the responsibility of persisting the book to the appropriate repository.